+++ inst/tinytest/test_biplot_ca.R
@@ -10,6 +10,10 @@ if (at_home()) {
   data("benthos")
   X <- ca(benthos)
 
+  # Symetric CA biplot
+  plot_biplot_symetric <- function() biplot(X, type = "symetric", labels = NULL)
+  expect_snapshot_plot(plot_biplot_symetric, "plot_biplot_symetric")
+
   # Row principal CA biplot
   plot_biplot_row <- function() biplot(X, type = "row", labels = NULL)
   expect_snapshot_plot(plot_biplot_row, "CA_biplot_row")